add json component
