+++ Common/Types/Workflow/ComponentID.ts
@@ -4,6 +4,9 @@ enum ComponentID {
     Schedule = 'schedule',
     JavaScriptCode = 'javascript',
     Manual = 'manual',
+    JsonToText = 'json-to-text',
+    TextToJson = 'text-to-json',
+    MergeJson = 'merge-json'
 }
 
 export default ComponentID;
