+++ CommonServer/Types/Workflow/Components/BaseModel/CreateOneBaseModel.ts
@@ -1,16 +1,15 @@
 import BaseModel from 'Common/Models/BaseModel';
 import BadDataException from 'Common/Types/Exception/BadDataException';
 import ComponentMetadata, { Port } from 'Common/Types/Workflow/Component';
-import DatabaseService from '../../../Services/DatabaseService';
-import ComponentCode, { RunOptions, RunReturnType } from '../ComponentCode';
+import DatabaseService from '../../../../Services/DatabaseService';
+import ComponentCode, { RunOptions, RunReturnType } from '../../ComponentCode';
 import BaseModelComponents from 'Common/Types/Workflow/Components/BaseModel';
 import Text from 'Common/Types/Text';
 import { JSONObject } from 'Common/Types/JSON';
-import Query from '../../Database/Query';
 import JSONFunctions from 'Common/Types/JSONFunctions';
-import Select from '../../Database/Select';
+import Exception from 'Common/Types/Exception/Exception';
 
-export default class FindOneBaseModel<
+export default class CreateOneBaseModel<
     TBaseModel extends BaseModel
 > extends ComponentCode {
     private modelService: DatabaseService<TBaseModel> | null = null;
@@ -25,14 +24,14 @@ export default class FindOneBaseModel<
                         i.id ===
                         `${Text.pascalCaseToDashes(
                             modelService.getModel().tableName!
-                        )}-find-one`
+                        )}-create-one`
                     );
                 }
             );
 
         if (!BaseModelComponent) {
             throw new BadDataException(
-                'Find one component for ' +
+                'Create one component for ' +
                     modelService.getModel().tableName +
                     ' not found.'
             );
@@ -74,71 +73,55 @@ export default class FindOneBaseModel<
                 );
             }
 
-            if (!args['query']) {
+            if (!args['json']) {
                 throw options.onError(
-                    new BadDataException('Query is undefined.')
+                    new BadDataException('JSON is undefined.')
                 );
             }
 
-            if (typeof args['query'] === 'string') {
-                args['query'] = JSON.parse(args['query'] as string);
+            if (typeof args['json'] === 'string') {
+                args['json'] = JSON.parse(args['json'] as string);
             }
 
-            if (typeof args['query'] !== 'object') {
+            if (typeof args['json'] !== 'object') {
                 throw options.onError(
-                    new BadDataException('Query is should be of type object.')
+                    new BadDataException('JSON is should be of type object.')
                 );
             }
 
             if (this.modelService.getModel().getTenantColumn()) {
-                (args['query'] as JSONObject)[
+                (args['json'] as JSONObject)[
                     this.modelService.getModel().getTenantColumn() as string
                 ] = options.projectId;
             }
 
-            if (!args['select']) {
-                throw options.onError(
-                    new BadDataException('Select Fields is undefined.')
-                );
-            }
-
-            if (typeof args['select'] === 'string') {
-                args['select'] = JSON.parse(args['select'] as string);
-            }
-
-            if (typeof args['select'] !== 'object') {
-                throw options.onError(
-                    new BadDataException(
-                        'Select Fields is should be of type object.'
-                    )
-                );
-            }
-
-            const model: TBaseModel | null = await this.modelService.findOneBy({
-                query: (args['query'] as Query<TBaseModel>) || {},
-                select: args['select'] as Select<TBaseModel>,
+            const model: TBaseModel = (await this.modelService.create({
+                data: JSONFunctions.fromJSON<TBaseModel>(
+                    (args['json'] as JSONObject) || {},
+                    this.modelService.entityType
+                ) as TBaseModel,
                 props: {
                     isRoot: true,
                 },
-            });
+            })) as TBaseModel;
 
             return {
                 returnValues: {
-                    model: model
-                        ? JSONFunctions.toJSON(
-                              model,
-                              this.modelService.entityType
-                          )
-                        : null,
+                    model: JSONFunctions.toJSON(
+                        model,
+                        this.modelService.entityType
+                    ),
                 },
                 executePort: successPort,
             };
         } catch (err: any) {
-            options.log('Error runnning component');
-
-            options.log(
-                err.message ? err.message : JSON.stringify(err, null, 2)
-            );
+            if (err instanceof Exception) {
+                options.log(err.getMessage());
+            } else {
+                options.log(
+                    err.message ? err.message : JSON.stringify(err, null, 2)
+                );
+            }
 
             return {
                 returnValues: {},
