+++ CommonServer/Types/Workflow/Components/BaseModel/DeleteManyBaseMoidel.ts
@@ -1,14 +1,16 @@
 import BaseModel from 'Common/Models/BaseModel';
 import BadDataException from 'Common/Types/Exception/BadDataException';
 import ComponentMetadata, { Port } from 'Common/Types/Workflow/Component';
-import DatabaseService from '../../../Services/DatabaseService';
-import ComponentCode, { RunOptions, RunReturnType } from '../ComponentCode';
+import DatabaseService from '../../../../Services/DatabaseService';
+import ComponentCode, { RunOptions, RunReturnType } from '../../ComponentCode';
 import BaseModelComponents from 'Common/Types/Workflow/Components/BaseModel';
 import Text from 'Common/Types/Text';
 import { JSONObject } from 'Common/Types/JSON';
-import Query from '../../Database/Query';
+import Query from '../../../Database/Query';
+import { LIMIT_PER_PROJECT } from 'Common/Types/Database/LimitMax';
+import PositiveNumber from 'Common/Types/PositiveNumber';
 
-export default class DeleteOneBaseModel<
+export default class DeleteManyBaseModel<
     TBaseModel extends BaseModel
 > extends ComponentCode {
     private modelService: DatabaseService<TBaseModel> | null = null;
@@ -23,14 +25,14 @@ export default class DeleteOneBaseModel<
                         i.id ===
                         `${Text.pascalCaseToDashes(
                             modelService.getModel().tableName!
-                        )}-delete-one`
+                        )}-delete-many`
                     );
                 }
             );
 
         if (!BaseModelComponent) {
             throw new BadDataException(
-                'Delete one component for ' +
+                'Delete many component for ' +
                     modelService.getModel().tableName +
                     ' not found.'
             );
@@ -88,14 +90,41 @@ export default class DeleteOneBaseModel<
                 );
             }
 
+            if (args['skip'] && typeof args['skip'] === 'string') {
+                args['skip'] = parseInt(args['skip']);
+            }
+
+            if (args['limit'] && typeof args['limit'] === 'string') {
+                args['limit'] = parseInt(args['limit']);
+            }
+
+            if (typeof args['skip'] !== 'number') {
+                args['skip'] = 0;
+            }
+
+            if (typeof args['limit'] !== 'number') {
+                args['limit'] = 10;
+            }
+
+            if (
+                typeof args['limit'] === 'number' &&
+                args['limit'] > LIMIT_PER_PROJECT
+            ) {
+                options.log('Limit cannot be ' + args['limit']);
+                options.log('Setting the limit to ' + LIMIT_PER_PROJECT);
+                args['limit'] = LIMIT_PER_PROJECT;
+            }
+
             if (this.modelService.getModel().getTenantColumn()) {
                 (args['query'] as JSONObject)[
                     this.modelService.getModel().getTenantColumn() as string
                 ] = options.projectId;
             }
 
-            await this.modelService.deleteOneBy({
+            await this.modelService.deleteBy({
                 query: (args['query'] as Query<TBaseModel>) || {},
+                limit: new PositiveNumber(args['limit'] as number),
+                skip: new PositiveNumber(args['skip'] as number),
                 props: {
                     isRoot: true,
                 },
