+++ CommonServer/Types/Workflow/Components/BaseModel/OnCreateBaseModel.ts
@@ -1,11 +1,11 @@
 import BaseModel from 'Common/Models/BaseModel';
-import DatabaseService from '../../../Services/DatabaseService';
+import DatabaseService from '../../../../Services/DatabaseService';
 import OnTriggerBaseModel from './OnTriggerBaseModel';
 
-export default class OnDeleteBaseModel<
+export default class OnCreateBaseModel<
     TBaseModel extends BaseModel
 > extends OnTriggerBaseModel<TBaseModel> {
     public constructor(modelService: DatabaseService<TBaseModel>) {
-        super(modelService, 'on-delete');
+        super(modelService, 'on-create');
     }
 }
