+++ CommonServer/Types/Workflow/Components/BaseModel/OnDeleteBaseModel.ts
@@ -1,11 +1,11 @@
 import BaseModel from 'Common/Models/BaseModel';
-import DatabaseService from '../../../Services/DatabaseService';
+import DatabaseService from '../../../../Services/DatabaseService';
 import OnTriggerBaseModel from './OnTriggerBaseModel';
 
-export default class OnCreateBaseModel<
+export default class OnDeleteBaseModel<
     TBaseModel extends BaseModel
 > extends OnTriggerBaseModel<TBaseModel> {
     public constructor(modelService: DatabaseService<TBaseModel>) {
-        super(modelService, 'on-create');
+        super(modelService, 'on-delete');
     }
 }
